template version test case
